src/application/requirements.txt
fastapi
uvicorn
python-dotenv
livekit-api
pydantic
